Add files via upload
An update to a lot of the systems. Working currently on the XMLWriter/CSVWriter/ and SqlDBConnection. The biggest change is in the autosetusername to get rid of any reference to the Window.h header file. I also need to do a check for the windows operating system so that we can use the arp system as arp -n doesnt seem to apply with windows